TASK 2: As a user I can mark a todo as completed.
Added a status field to table todos.  Added a migration SQL file and
updated the original create table statement.  Replaced INT(11) with
INTEGER for the sake of consitency.

Added another task to main.py to run the migration.  This does nothing
more than feed the sql in a given migration file to sqlite but anything
more complicated would have exceeded the scope of the exercise.

Changed todo_delete() to todo_post() and extended it to handle deletion
and toggling the status.

Updated the templates to handle the new functionality.

Fixed multiple bugs in sqliteorm.